+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+## [0.1.5] - 2021-12-??
+### Changed
+-   Refactor : fetching data script in a separate file (api.js).
+
+### Added
+-   Display photographer media.
+
 ## [0.1.4] - 2021-12-13
 ### Changed
 -   Use a single "index.js" to import all scripts.